movement

+++ src/be/uantwerpen/fti/ei/J2D/data_oriented/J2DVisualComp.java
@@ -1,5 +1,6 @@
 package be.uantwerpen.fti.ei.J2D.data_oriented;
 
+import be.uantwerpen.fti.ei.IntPtr;
 import be.uantwerpen.fti.ei.J2D.GraphicsContext;
 import be.uantwerpen.fti.ei.components.AVisualComp;
 
@@ -8,7 +9,7 @@ import java.awt.*;
 public class J2DVisualComp extends AVisualComp {
     GraphicsContext grCtx;
     int r, g, b;
-    public J2DVisualComp(int x, int y, int size, int r, int g, int b, GraphicsContext grCtx) {
+    public J2DVisualComp(IntPtr x, IntPtr y, int size, int r, int g, int b, GraphicsContext grCtx) {
         super(x, y, size);
         this.r = r;
         this.g = g;
